1、修复员工登录更新状态bug 2、返回配置信息bug

+++ ruoyi-system/src/main/java/com/ruoyi/system/fantang/service/IFtStaffDemandDaoService.java
@@ -12,7 +12,7 @@ import com.baomidou.mybatisplus.extension.service.IService;
  */
 public interface IFtStaffDemandDaoService extends IService<FtStaffDemandDao> {
 
-    AjaxResult getConfiguration(Integer staffId);
+    AjaxResult getConfiguration(Long staffId);
 
     AjaxResult setDemandMode(Long id, Boolean demandModeFlag);
 }
